subsets
When the function search is called with a parameter k, it decides whether to include the element k in the subset or not, and in both cases, then calls itself with parameter k + 1. then if k = n + 1, then the function notices all the elements have been processed and a subset have been generated.